+++ History.md
@@ -4,3 +4,8 @@
 * Add(): 增加Tcp客户端；增加Tcp服务端
 * Add(): 简单状态机
 * Add(): ZMQ的C#测试Demo
+* Add(): ZMQ: TaskVent、TaskWorker、TaskSink方法测试分布式处理
+	-- 先开启TaskSink
+	-- 然后TaskVent
+	-- 最后开启一个或多个TaskWorker
+	-- 结论： 开启一个worker用时5039ms，而开启2个用时1700，判别为并行处理模型